refactor(ui): extract trivial tabs (advancements, settings, about)
Three small per-tab mixins under l5r/ui/tabs/:

* AdvancementsTabMixin -> Tab 6 (build_ui_page_6, refund button +
  advancement history table).
* SettingsTabMixin -> Tab 11 (build_ui_page_settings, wraps the
  existing widgets.SettingsWidget).
* AboutTabMixin -> Tab 12 (build_ui_page_about, version/links/credits
  page; imports its constants explicitly from l5r.l5rcmcore now that
  it no longer rides on the main.py star-import).

main.py: 1755 -> 1674 lines.
